requirements.txt
discord.py~=1.5.1
pyyaml~=5.3.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
asyncio~=3.4.3
python-dateutil~=2.8.1
pytz~=2020.4
google~=3.0.0
html2text~=2020.1.16
feedparser~=6.0.2
